Issue-212: Added RecyclerViewExamplesActivity.
