+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,10 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name=".RecyclerViewExamplesActivity"
+            android:label="@string/recycler_view_examples" />
+
     </application>
 
 </manifest>
