+++ app/src/main/java/com/example/kirchhoff/example/MainActivity.kt
@@ -1,36 +1,27 @@
 package com.example.kirchhoff.example
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import com.example.bottomnavigationview.BottomNavigationViewActivity
 import com.example.constraintanimation.ConstraintLayoutAnimationActivity
 import com.example.custommenuitem.CustomMenuItemActivity
+import com.example.kirchhoff.example.extension.startActivity
 import com.example.lazybind.LazyBindActivity
 import com.example.likebutton.LikeButtonActivity
 import com.example.prefixedittext.PrefixEditTextActivity
 import com.example.progressbarcustomization.ProgressBarCustomizationActivity
 import com.example.raindropview.RainDropActivity
-import com.example.recyclercollapsing.RecyclerCollapsingToolbarActivity
-import com.example.recyclerdragdrop.RecyclerViewDragDropActivity
-import com.example.recyclersmoothswipe.RecyclerSmoothSwipeActivity
-import com.example.recyclertoviewpager.RecyclerToViewPagerActivity
 import com.example.screenshot.ScreenshotActivity
 import com.example.simpletransition.SimpleTransitionActivity
-import com.example.timediffutils.TimeDiffUtilActivity
 import com.kirchhoff.activitytoservice.ActivityToService
 import com.kirchhoff.circularrevealfragment.CircularFragmentAnimationActivity
 import com.kirchhoff.coordinatebehavior.CoordinatorBehaviorActivity
 import com.kirchhoff.countdownlabel.CountdownLabelActivity
 import com.kirchhoff.curvedbottombar.CurvedBottomBarActivity
 import com.kirchhoff.fabanimation.FabAnimationActivity
-import com.kirchhoff.fastrecyclerview.FastRecyclerViewActivity
 import com.kirchhoff.fragmentcommunication.FragmentCommunicationActivity
 import com.kirchhoff.gmailselectinganimation.GmailSelectionActivity
-import com.kirchhoff.recyclermvp.MvpRecyclerViewActivity
-import com.kirchhoff.recyclersearch.RecyclerSearchActivity
-import com.kirchhoff.thingdiffutils.ThingDiffUtilActivity
 import com.kirchhoff.viewpagerwithoutfragment.VPWithoutFragmentActivity
 
 @Suppress("UNUSED_PARAMETER")
@@ -41,24 +32,16 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.a_main)
     }
 
-    fun recyclerViewWithSearch(view: View) {
-        startActivity(RecyclerSearchActivity::class.java)
+    fun recyclerViewExamples(view: View) {
+        startActivity(RecyclerViewExamplesActivity::class.java)
     }
 
     fun fabAnimation(view: View) {
         startActivity(FabAnimationActivity::class.java)
     }
 
-    fun recyclerCollapsing(view: View) {
-        startActivity(RecyclerCollapsingToolbarActivity::class.java)
-    }
-
     fun vpWithoutFragment(view: View) {
         startActivity(VPWithoutFragmentActivity::class.java)
-    }
-
-    fun timeDiffUtil(view: View) {
-        startActivity(TimeDiffUtilActivity::class.java)
     }
 
     fun likeButtonClick(view: View) {
@@ -75,18 +58,6 @@ class MainActivity : AppCompatActivity() {
 
     fun customCoordinateLayoutBehavior(view: View) {
         startActivity(CoordinatorBehaviorActivity::class.java)
-    }
-
-    fun mvpRecycler(view: View) {
-        startActivity(MvpRecyclerViewActivity::class.java)
-    }
-
-    fun thingDiffUtils(view: View) {
-        startActivity(ThingDiffUtilActivity::class.java)
-    }
-
-    fun fastRecyclerViewExample(view: View) {
-        startActivity(FastRecyclerViewActivity::class.java)
     }
 
     fun curvedBottomBarExample(view: View) {
@@ -109,20 +80,8 @@ class MainActivity : AppCompatActivity() {
         startActivity(RainDropActivity::class.java)
     }
 
-    fun recyclerSmoothSwipe(view: View) {
-        startActivity(RecyclerSmoothSwipeActivity::class.java)
-    }
-
     fun constraintLayoutAnimation(view: View) {
         startActivity(ConstraintLayoutAnimationActivity::class.java)
-    }
-
-    fun recyclerDragDrop(view: View) {
-        startActivity(RecyclerViewDragDropActivity::class.java)
-    }
-
-    fun recyclerToViewPager(view: View) {
-        startActivity(RecyclerToViewPagerActivity::class.java)
     }
 
     fun prefixEditText(view: View) {
@@ -151,9 +110,5 @@ class MainActivity : AppCompatActivity() {
 
     fun customMenuItem(view: View) {
         startActivity(CustomMenuItemActivity::class.java)
-    }
-
-    private fun startActivity(cls: Class<*>) {
-        startActivity(Intent(this, cls))
     }
 }
