+++ app/src/main/java/com/example/kirchhoff/example/extension/Activity.kt
@@ -1,0 +1,8 @@
+package com.example.kirchhoff.example.extension
+
+import android.app.Activity
+import android.content.Intent
+
+fun Activity.startActivity(cls: Class<*>) {
+    startActivity(Intent(this, cls))
+}
