+++ app/src/main/res/layout/a_main.xml
@@ -12,11 +12,10 @@
         android:orientation="vertical">
 
         <Button
-            android:id="@+id/recyclerViewWithSearch"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:onClick="recyclerViewWithSearch"
-            android:text="@string/recyclerview_with_search_example" />
+            android:onClick="recyclerViewExamples"
+            android:text="@string/recycler_view_examples" />
 
         <Button
             android:id="@+id/fabAnimationExample"
@@ -26,25 +25,11 @@
             android:text="@string/fab_animation_example" />
 
         <Button
-            android:id="@+id/recyclerCollapsing"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="recyclerCollapsing"
-            android:text="@string/recycler_collapsing_example" />
-
-        <Button
             android:id="@+id/vpWithoutFragment"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="vpWithoutFragment"
             android:text="@string/vp_without_fragment_example" />
-
-        <Button
-            android:id="@+id/timeDiffUtil"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="timeDiffUtil"
-            android:text="@string/time_diff_utils_example" />
 
         <Button
             android:id="@+id/likeButtonClick"
@@ -73,27 +58,6 @@
             android:layout_height="wrap_content"
             android:onClick="customCoordinateLayoutBehavior"
             android:text="@string/coordinator_layout_behavior_title" />
-
-        <Button
-            android:id="@+id/mvpRecycler"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="mvpRecycler"
-            android:text="@string/mvp_recyclerview_example" />
-
-        <Button
-            android:id="@+id/thingDiffUtils"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="thingDiffUtils"
-            android:text="@string/thing_diffutils_example" />
-
-        <Button
-            android:id="@+id/fastRecyclerViewExample"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="fastRecyclerViewExample"
-            android:text="@string/fast_recycler_view" />
 
         <Button
             android:id="@+id/curvedBottomBarExample"
@@ -131,32 +95,11 @@
             android:text="@string/raindrop_view_example" />
 
         <Button
-            android:id="@+id/bRecyclerSmoothSwipe"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="recyclerSmoothSwipe"
-            android:text="@string/recycler_smooth_swipe_example" />
-
-        <Button
             android:id="@+id/bConstraintLayoutAnimation"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="constraintLayoutAnimation"
             android:text="@string/constraint_layout_animation_example" />
-
-        <Button
-            android:id="@+id/bRecyclerDragDrop"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="recyclerDragDrop"
-            android:text="@string/recycler_view_drag_and_drop_example" />
-
-        <Button
-            android:id="@+id/bRecyclerToViewPager"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="recyclerToViewPager"
-            android:text="@string/recycler_to_viewpager_title" />
 
         <Button
             android:id="@+id/bPrefixEditText"
