+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Example</string>
+    <string name="recycler_view_examples">RecyclerView examples</string>
 </resources>
